--- src/main/java/com/reservation/reserve/event/ReservationCreatedEvent.java
@@ -1,13 +1,0 @@
-package com.reservation.reserve.event;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class ReservationCreatedEvent {
-    private Long reservationId;
-    private String email;
-    private Long concertId;
-    private Long seatId;
-}
