+++ src/main/java/com/reservation/reserve/event/ReservationEventDto.java
@@ -1,0 +1,14 @@
+package com.reservation.reserve.event;
+
+import java.time.LocalDateTime;
+
+
+public record ReservationEventDto(
+        Long reservationId,
+        String email,
+        String concertName,
+        LocalDateTime concertDate,
+        String seatNumber,
+        String eventType,
+        String timestamp
+) { }
